# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Homeroom API
  version: 0.1.0
  description: |
    Contract shared by the Next.js web app and the Flutter app. Every route is scoped to the
    caller's school and role by Postgres row-level security (see backend/migrations); the
    "Roles" line on each operation says who is meant to call it, and the database enforces it.

    **Conventions**
    - Auth: `Authorization: Bearer <jwt>`. The API maps the token to `people.id` and sets `app.person_id`.
    - IDs are UUIDs. Dates are `YYYY-MM-DD`; timestamps are RFC 3339 UTC. Money is a decimal string in GHS minor-unit-safe form (`"1850.00"`), never a float.
    - Lists are cursor-paginated: `?limit=50&cursor=...` returns `{ items, next_cursor }`.
    - Errors use `application/problem+json`. A row the caller may not see is reported as **404**, not 403, so existence is not leaked. 403 is for a visible resource the caller may not change.
    - Writes from the offline apps send an `Idempotency-Key` header; replays return the original result.
servers:
  - url: https://api.homeroom.example/v1
  - url: http://localhost:3000/api/v1
security:
  - bearer: []
...
paths:
...
  /terms:
    get:
      operationId: listTerms
      tags: [ Classes ]
      summary: Terms for the school
      description: 'Roles: any.'
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Terms
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Term' } } }, required: [ items ] } } }
    post:
      operationId: createTerm
      tags: [ Admin ]
      summary: Create a term
      description: 'Roles: admin.'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TermInput' } } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '201': { description: 'Created', content: { application/json: { schema: { $ref: '#/components/schemas/Term' } } } }
  /terms/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    patch:
      operationId: updateTerm
      tags: [ Admin ]
      summary: 'Edit a term, or close it'
      description: 'Roles: admin. `closed: true` locks teacher edits for every section in the term.'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TermInput' } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Updated', content: { application/json: { schema: { $ref: '#/components/schemas/Term' } } } }
  /class_sections:
    get:
      operationId: listClassSections
      tags: [ Classes ]
      summary: Class sections the caller can see
      description: 'Roles: admin (all), teacher (taught), guardian (children''s).'
      parameters: [ { name: term_id, in: query, schema: { type: string, format: uuid } } ]
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Sections
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/ClassSection' } } }, required: [ items ] } } }
    post:
      operationId: createClassSection
      tags: [ Admin ]
      summary: Create a class section
      description: 'Roles: admin.'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClassSectionInput' } } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '201': { description: 'Created', content: { application/json: { schema: { $ref: '#/components/schemas/ClassSection' } } } }
  /class_sections/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      operationId: getClassSection
      tags: [ Classes ]
      summary: 'Section with roster, teachers, attendance % and running average'
      description: 'Roles: admin, teacher (taught).'
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Section', content: { application/json: { schema: { $ref: '#/components/schemas/ClassSectionDetail' } } } }
    patch:
      operationId: updateClassSection
      tags: [ Admin ]
      summary: 'Edit a section, e.g. change the homeroom teacher'
      description: 'Roles: admin.'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClassSectionInput' } } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Updated', content: { application/json: { schema: { $ref: '#/components/schemas/ClassSection' } } } }
  /class_sections/{id}/periods:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      operationId: getTimetable
      tags: [ Classes ]
      summary: Weekly timetable for a section
      description: 'Roles: admin, teacher, guardian (child''s section).'
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Periods
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Period' } } }, required: [ items ] } } }
    put:
      operationId: setTimetable
      tags: [ Admin ]
      summary: Replace the section's timetable
      description: 'Roles: admin.'
      requestBody:
        required: true
        content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PeriodInput' } } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Saved timetable', content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Period' } } }, required: [ items ] } } } }
  /enrollments:
    post:
      operationId: enrollStudent
      tags: [ Admin ]
      summary: Assign a student to a class section
      description: 'Roles: admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ student_id, class_section_id ]
              properties: { student_id: { type: string, format: uuid }, class_section_id: { type: string, format: uuid } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '201': { description: 'Enrolled', content: { application/json: { schema: { $ref: '#/components/schemas/Enrollment' } } } }
        '409': { $ref: '#/components/responses/Problem' }
  /enrollments/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    delete:
      operationId: withdrawStudent
      tags: [ Admin ]
      summary: Withdraw a student from a section
      description: 'Roles: admin. **Soft delete**: sets status to `withdrawn`; the row and its history stay for transcripts (the database has no DELETE policy).'
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '204': { description: 'Withdrawn' }
  /class_sections/{id}/students/import:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      operationId: importStudents
      tags: [ Admin ]
      summary: Bulk-create students and enrol them in this section
      description: 'Roles: admin. Each row becomes a new student (there is no matching against existing people); build the array from a CSV client-side, one row per pupil.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ students ]
              properties:
                students:
                  type: array
                  minItems: 1
                  maxItems: 300
                  items:
                    type: object
                    required: [ full_name ]
                    properties: { full_name: { type: string, maxLength: 120 } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '201':
          description: Created and enrolled
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/StudentRef' } } }, required: [ items ] } } }

  # ------------------------------------------------------------------ attendance
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [ status, title ]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }
        code: { type: string, example: term_closed }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
        request_id:
          type: string
          description: Matches the X-Request-Id response header and the server logs
...
    Role: { type: string, enum: [ admin, teacher, guardian, student ] }
...
    StudentRef:
      type: object
      required: [ id, full_name ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
    Student:
      allOf:
        - $ref: '#/components/schemas/StudentRef'
        - type: object
          properties:
            grade_level: { type: string }
            attendance_rate_pct: { type: number, nullable: true }
            guardian_names: { type: array, items: { type: string }, description: 'Admin lists only: primary contact first' }
          required: [ attendance_rate_pct ]
...
    Person:
      type: object
      required: [ id, full_name, role ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        contact: { type: object, properties: { phone: { type: string }, email: { type: string } } }
        assignments:
          type: array
          description: 'Teachers, in the admin directory: what they teach, in the current term'
          items:
            type: object
            properties:
              class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
              subject: { type: string }
            required: [ class_section, subject ]
...
    Term:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        closed: { type: boolean }
      required: [ id, name, starts_on, ends_on, closed ]
    TermInput:
      type: object
      properties:
        name: { type: string }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        closed: { type: boolean }
    ClassSection:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 6A }
        grade_level: { type: string }
        term_id: { type: string, format: uuid }
        homeroom_teacher: { $ref: '#/components/schemas/Person' }
        subject: { type: string, description: 'Present on /me and /class_sections for a teacher: the subject they teach here' }
      required: [ id, name, grade_level, term_id ]
    ClassSectionInput:
      type: object
      properties:
        name: { type: string }
        grade_level: { type: string }
        term_id: { type: string, format: uuid }
        homeroom_teacher_id: { type: string, format: uuid, nullable: true }
    ClassSectionDetail:
      allOf:
        - $ref: '#/components/schemas/ClassSection'
        - type: object
          properties:
            teachers:
              type: array
              items: { type: object, properties: { person: { $ref: '#/components/schemas/Person' }, subject: { type: string } }, required: [ person, subject ] }
            roster: { type: array, items: { $ref: '#/components/schemas/Student' } }
            attendance_rate_pct: { type: number, nullable: true }
          required: [ teachers, roster, attendance_rate_pct ]
    Enrollment:
      type: object
      properties:
        id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        class_section_id: { type: string, format: uuid }
        class_section_name: { type: string }
        term_name: { type: string }
        status: { type: string, enum: [ active, completed, withdrawn ] }
      required: [ id, student_id, class_section_id, status ]
    Period:
      type: object
      properties:
        id: { type: string, format: uuid }
        subject: { type: string }
        teacher: { $ref: '#/components/schemas/Person' }
        weekday: { type: integer, minimum: 1, maximum: 7, description: 'ISO, 1 = Monday' }
        starts_at: { type: string, example: '08:00' }
        ends_at: { type: string, example: '08:45' }
      required: [ id, subject, teacher, weekday, starts_at, ends_at ]
    PeriodInput:
      type: object
      required: [ subject, teacher_id, weekday, starts_at, ends_at ]
      properties:
        subject: { type: string }
        teacher_id: { type: string, format: uuid }
        weekday: { type: integer }
        starts_at: { type: string }
        ends_at: { type: string }
...